+++ src/gwt/src/org/rstudio/studio/client/workbench/views/vcs/git/dialog/GitReviewPanel.java
@@ -39,9 +39,11 @@ import com.google.gwt.user.client.ui.*;
 import com.google.gwt.user.client.ui.PopupPanel.PositionCallback;
 import com.google.inject.Inject;
 
+import org.rstudio.core.client.BrowseCap;
 import org.rstudio.core.client.WidgetHandlerRegistration;
 import org.rstudio.core.client.command.AppCommand;
 import org.rstudio.core.client.command.KeyboardShortcut;
+import org.rstudio.core.client.dom.DomUtils;
 import org.rstudio.core.client.widget.*;
 import org.rstudio.studio.client.common.filetypes.FileTypeRegistry;
 import org.rstudio.studio.client.common.vcs.GitServerOperations.PatchMode;
@@ -310,7 +312,6 @@ public class GitReviewPanel extends ResizeComposite implements Display
       listBoxAdapter_ = new ListBoxAdapter(contextLines_);
 
       FontSizer.applyNormalFontSize(commitMessage_);
-
       new WidgetHandlerRegistration(this)
       {
          @Override
@@ -322,10 +323,11 @@ public class GitReviewPanel extends ResizeComposite implements Display
                public void onPreviewNativeEvent(NativePreviewEvent event)
                {
                   NativeEvent nativeEvent = event.getNativeEvent();
-                  if (event.getTypeInt() == Event.ONKEYDOWN
-                      && KeyboardShortcut.getModifierValue(nativeEvent) == KeyboardShortcut.CTRL)
+                  int keyCode = nativeEvent.getKeyCode();
+                  int modifier = KeyboardShortcut.getModifierValue(nativeEvent);
+                  if (event.getTypeInt() == Event.ONKEYDOWN && isCtrlOrMeta(modifier))
                   {
-                     switch (nativeEvent.getKeyCode())
+                     switch (keyCode)
                      {
                         case KeyCodes.KEY_DOWN:
                            nativeEvent.preventDefault();
@@ -343,12 +345,26 @@ public class GitReviewPanel extends ResizeComposite implements Display
                            nativeEvent.preventDefault();
                            scrollBy(diffScroll_, -getPageScroll(diffScroll_), 0);
                            break;
+                        case KeyCodes.KEY_ENTER:
+                           if (DomUtils.getActiveElement() == commitMessage_.getElement())
+                           {
+                              nativeEvent.preventDefault();
+                              commitButton_.click();
+                           }
+                           break;
                      }
                   }
                }
             });
          }
       };
+   }
+   
+   private boolean isCtrlOrMeta(int modifier)
+   {
+      return BrowseCap.isMacintosh()
+            ? (modifier == KeyboardShortcut.CTRL || modifier == KeyboardShortcut.META)
+            : modifier == KeyboardShortcut.CTRL;
    }
 
    private void scrollBy(ScrollPanel scrollPanel, int vscroll, int hscroll)